backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
py-ptsl==500.1.1
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
protobuf>=5.26.1,<5.27.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
pytest==7.4.3
pytest-asyncio==0.21.1
loguru==0.7.2
python-dotenv==1.0.0